requirements.txt
bcrypt
fuzzywuzzy
python-Levenshtein
flask
flask-sqlalchemy
gunicorn
psycopg2==2.6.1
Flask-Migrate==1.8.0
requests
Cython
pandas
python-resize-image
image
flask-mysql
mysql-python
